Automated behavioral tests can now be run either on a host on which the IDA service is installed, in which case all tests are run, or against a remote IDA service instance, in which case a subset of tests are run, limited to tests which only require user credentials. This enables testing in all environments from which users are likely to be using the command line tools.
All test data which is uploaded to the configured IDA service instance by the automated tests is put into a timestamped and token unique folder to avoid collision with existing data or other tests.

Support is provided for netrc for secure configuration of credentials. The same configuration used for the command line tools also can be utilized for executing tests using only user credentials.