File uploads are now conditional such that if a file already exists at the specified target relative path and is the same size as the local file to be uploaded, no upload will occur. This enables both easy resumption of interrupted uploads as well as repeated uploads of new files in a particular folder scope.
An sha256 checksum is now generated for each file that is uploaded, which is provided to Nextcloud for storage in the file cache record, for later inclusion in file details reported by the 'info' command, and which can be used later for validation of the upload once support is added to the IDA API.

Added a new 'validate' action which can be used to check whether local files exist in IDA and have the same size in bytes.

Added a new command line parameter -F for forcing upload of files regardless of whether they exist in IDA.

Added a new command line parameter -D for executing dry-run such that operations with side effects will be skipped while verbose output informing which operations would be executed if not a dry-run.

Added a new command line parameter -j for obtaining info action output formatted as JSON rather than plain text with indentation.

All verbose and debug statements are now output to stderr, providing clean separation from info and inventory output.